requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.6.0
anyio==4.7.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==3.0.0
async-lru==2.0.4
attrs==24.3.0
babel==2.16.0
beautifulsoup4==4.13.4
bleach==6.2.0
brotlicffi==1.0.9.2
certifi==2025.10.5
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.8
colorama==0.4.6
comm==0.2.1
contourpy==1.3.1
cycler==0.11.0
debugpy==1.8.11
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
executing==0.8.3
fastjsonschema==2.20.0
fonttools==4.60.1
frozenlist==1.7.0
greenlet==3.1.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
idna==3.7
ipykernel==6.29.5
ipython==9.1.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.6.1
json5==0.9.25
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2023.7.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kiwisolver==1.4.8
langchain==0.3.27
langchain-core==0.3.72
langchain-mcp-adapters==0.1.9
langchain-ollama==0.3.6
langchain-openai==0.3.16
langchain-tavily==0.2.11
langchain-text-splitters==0.3.11
langgraph==0.6.3
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.3
langgraph-sdk==0.2.0
langsmith==0.3.45
MarkupSafe==3.0.2
matplotlib==3.10.6
matplotlib-inline==0.1.6
mcp==1.13.1
mistune==3.1.2
mkl_fft==2.1.1
mkl_random==1.3.0
mkl-service==2.5.2
multidict==6.6.4
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest_asyncio==1.6.0
notebook==7.4.4
notebook_shim==0.2.4
numpy==2.3.3
ollama==0.5.1
openai==1.77.0
orjson==3.10.14
ormsgpack==1.10.0
overrides==7.4.0
packaging==24.2
pandocfilters==1.5.0
parso==0.8.4
pillow==11.3.0
pip==25.1
platformdirs==4.3.7
prometheus_client==0.21.1
prompt_toolkit==3.0.43
propcache==0.3.2
psutil==5.9.0
pure-eval==0.2.2
pycparser==2.21
pydantic==2.11.7
pydantic_core==2.33.2
pydantic-settings==2.10.1
Pygments==2.19.1
pyparsing==3.2.0
PyQt6==6.9.1
PyQt6_sip==13.10.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-json-logger==3.2.1
python-multipart==0.0.20
pywin32==311
pywinpty==2.0.15
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.30.2
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
rfc3339_validator==0.1.4
rfc3986_validator==0.1.1
rpds-py==0.22.3
Send2Trash==1.8.2
setuptools==78.1.1
sip==6.12.0
six==1.17.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.41
sse-starlette==3.0.2
stack-data==0.2.0
starlette==0.47.3
tenacity==9.0.0
terminado==0.17.1
tiktoken==0.9.0
tinycss2==1.4.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
typing-inspection==0.4.0
urllib3==2.5.0
uvicorn==0.35.0
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
wheel==0.45.1
win_inet_pton==1.1.0
xxhash==3.5.0
yarl==1.20.1
zstandard==0.23.0
